@@ -1,5 +1,3 @@
 import FetchData from './FetchData';
 
-export {
-  FetchData
-};
+export { FetchData };
