@@ -1,5 +1,3 @@
 import Form from './Form';
 
-export {
-  Form
-};
+export { Form };
