@@ -1,5 +1,3 @@
 import Login from './Login';
 
-export {
-  Login
-};
+export { Login };
