@@ -1,3 +1,20 @@
+﻿import { actionCreators } from './actions';
+
+const { increment, decrement, handleOnCheck, selectOption } = actionCreators;
+
+export type FormPayload =
+  | ReturnType<typeof increment>
+  | ReturnType<typeof decrement>
+  | ReturnType<typeof handleOnCheck>
+  | ReturnType<typeof selectOption>;
+
+export enum FormActionType {
+  INCREMENT = 'form/increment',
+  DECREMENT = 'form/decrement',
+  CHECKBOX_CHECK = 'form/checkbox_check',
+  DROPDOWN_SELECT = 'form/dropdown_select'
+};
+
 export type IDropdownOption = Readonly<{
   value: number;
   label: string;
@@ -8,10 +25,3 @@ export type IFormState = Readonly<{
   checkboxValue: boolean;
   selectedDropdownOption: IDropdownOption;
 }>;
-
-export enum FormActionType {
-  INCREMENT = 'form/increment',
-  DECREMENT = 'form/decrement',
-  CHECKBOX_CHECK = 'form/checkbox_check',
-  DROPDOWN_SELECT = 'form/dropdown_select'
-};
