@@ -1,12 +1,14 @@
-import { SampleApi } from '../../api';
-import { IAppThunkAction, ReduxAction } from '../';
-import { WeatherActionType, IWeatherForecast } from './types';
+﻿import { SampleApi } from '../../api';
+import { ReduxAction, AppThunk } from '../';
+import { WeatherActionType, IWeatherForecast, WeatherForecastPayload } from './types';
 
 export const actionCreators = {
-  resetState: (): ReduxAction => ({
-    type: WeatherActionType.RESET_STATE
+  resetState: (): ReduxAction<WeatherForecastPayload> => ({
+    type: WeatherActionType.RESET_STATE,
   }),
-  requestWeatherForecasts: (startDateIndex: number): IAppThunkAction<ReduxAction> => (dispatch, getState) => {
+  requestWeatherForecasts: (
+    startDateIndex: number
+  ): AppThunk<WeatherForecastPayload> => async (dispatch, getState) => {
     // If param startDateIndex === state.startDateIndex, do not perform action
     if (startDateIndex === getState().weatherForecasts.startDateIndex) {
       return;
@@ -14,17 +16,16 @@ export const actionCreators = {
 
     // Dispatch request
     dispatch({
-      startDateIndex,
-      type: WeatherActionType.REQUEST
+      payload: { startDateIndex },
+      type: WeatherActionType.REQUEST,
     });
 
     // Build http request and success handler in Promise<void> wrapper
     SampleApi.getWeatherForecastsAsync(startDateIndex)
       .then((forecasts: IWeatherForecast[]) => {
         dispatch({
-          forecasts,
-          startDateIndex,
-          type: WeatherActionType.RECEIVE
+          type: WeatherActionType.RECEIVE,
+          payload: { forecasts, startDateIndex },
         });
       });
   },
