@@ -1,5 +1,3 @@
-const createClassName = (classNames: any[]): string => classNames.filter(x => x).join(' ');
+const createClassName = (classNames: any[]): string => classNames.filter((x) => x).join(' ');
 
-export {
-  createClassName
-};
+export { createClassName };
