@@ -1,5 +1,3 @@
 const isArrayWithLength = (test: any): boolean => Array.isArray(test) && !!test.length;
 
-export {
-  isArrayWithLength
-};
+export { isArrayWithLength };
