@@ -1,5 +1,3 @@
 const isNullOrUndefined = (test: any): boolean => test === null || test === undefined;
 
-export {
-  isNullOrUndefined
-};
+export { isNullOrUndefined };
