@@ -1,4 +1,4 @@
-import React, { ReactNode } from 'react';
+import { ReactNode } from 'react';
 import styled from 'styled-components';
 import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
@@ -21,6 +21,4 @@ const renderToastifyMsg = (message: string, icon?: IconProp): ReactNode => (
   </ToastifyMsg>
 );
 
-export {
-  renderToastifyMsg
-};
+export { renderToastifyMsg };
