@@ -2,12 +2,14 @@
 using GhostUI.Models;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.SignalR;
 
 namespace GhostUI.Controllers
 {
+    [ApiController]
     [Route("api/[controller]/[action]")]
-    public class AuthController : Controller
+    public class AuthController : ControllerBase
     {
         private readonly IHubContext<UsersHub> _hubContext;
 
@@ -17,6 +19,7 @@ namespace GhostUI.Controllers
         }
 
         [HttpPost]
+        [ProducesResponseType(typeof(AuthUser), StatusCodes.Status200OK)]
         public async Task<IActionResult> Login([FromBody]Credentials request)
         {
             await _hubContext.Clients.All.SendAsync("UserLogin");
@@ -25,6 +28,7 @@ namespace GhostUI.Controllers
         }
 
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<IActionResult> Logout()
         {
             await _hubContext.Clients.All.SendAsync("UserLogout");
