@@ -7,18 +7,20 @@ using System.Collections.Immutable;
 
 namespace GhostUI.Controllers
 {
-    [Route("api/[controller]")]
-    public class SampleDataController : Controller
+    [ApiController]
+    [Route("api/[controller]/[action]")]
+    public class SampleDataController : ControllerBase
     {
         public static readonly ImmutableArray<string> Summaries = ImmutableArray.Create(new[]
         {
             "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
         });
 
-        [HttpGet("[action]")]
+        [HttpGet]
         public IEnumerable<WeatherForecast> WeatherForecasts(int startDateIndex)
         {
             var rng = new Random();
+
             return Enumerable.Range(1, 5).Select(index => new WeatherForecast
             {
                 DateFormatted = DateTime.Now.AddDays(index + startDateIndex).ToString("d"),
